requirements.txt
# Airflow
apache-airflow==2.10.0

# Bibliotecas necessárias
setuptools
requests==2.31.0
psycopg2-binary==2.9.7
pandas==2.2.3
numpy>=2.0
pendulum==3.0.0
python-dotenv